+++ src/util/constants.js
@@ -71,6 +71,10 @@ export const CUSTOM_HTML_PREVIEW = 'ctCustomHtmlPreview'
 export const WEB_POPUP_PREVIEW = 'ctWebPopupPreview'
 export const QUALIFIED_CAMPAIGNS = 'WZRK_QC'
 export const CUSTOM_CT_ID_PREFIX = '_w_'
+export const BLOCK_REQUEST_COOKIE = 'WZRK_BLOCK'
+
+// Flag key for optional sub-domain profile isolation
+export const ISOLATE_COOKIE = 'WZRK_ISOLATE_SD'
 
 export const WEB_NATIVE_TEMPLATES = {
   KV_PAIR: 1,